Extract native file for attachment, and fix(?) attachment size
We extract the native file associated with an attachment, and store it
in memory (to avoid needing file IO and unique filename generation in
the document class).

Note that the attachment.size() and attachment.get_bytes().size()
functions return different values.  I've asked upstream for
clarification:

  http://pstsdk.codeplex.com/Thread/View.aspx?ThreadId=217470

For now, based on some quick hand testing, I'm temporarily going with
the theory that get_bytes().size() is actually correct.  We will need to
follow this thread and update the code if this assumption turns out to
be wrong.