Make document constructors explicit
I just spent ~5 minutes trying to debug a problem caused by passing an
attachment directly to a function that wanted a document.  C++ was all
too happy to perform an implicit conversion using the matching document
constructor.

By marking these constructors as explicit, we prevent the implicit
conversion that confused me for a couple of minutes.  I really shouldn't
need ~5 minutes to figure this out.